[feature] Check "html-version" output property when creating "html" or "xhtml" serializer. For backwards compatibility, using the old methods "html5", "xhtml5" etc. is still possible.
